Update Xcode project files for 6.4 - Physics Rig

+++ TestRigs/PhysicsRig/PSArborTouchViewController.m
@@ -292,7 +292,7 @@
     self.sumLabel.text      = [NSString stringWithFormat:@"%f", _integrator.energy.sum];
     self.maxLabel.text      = [NSString stringWithFormat:@"%f", _integrator.energy.max];
     self.meanLabel.text     = [NSString stringWithFormat:@"%f", _integrator.energy.mean];
-    self.countLabel.text    = [NSString stringWithFormat:@"%u", _integrator.energy.count];
+    self.countLabel.text    = [NSString stringWithFormat:@"%lu", (unsigned long)_integrator.energy.count];
     
     self.p1Name.text        = _particle1.name;
     self.p1Mass.text        = [NSString stringWithFormat:@"%f", _particle1.mass];
@@ -310,7 +310,7 @@
     self.particleView4.center = [self toScreen:_particle4.position];
     
     _counter++;
-    self.counterLabel.text = [NSString stringWithFormat:@"%d", _counter];
+    self.counterLabel.text = [NSString stringWithFormat:@"%ld", (long)_counter];
     
     [_viewPort setNeedsDisplay];
 }
